refactor the experiment code
