+++ library/src/test/experiment/relative_entroypy_detector_evaluation.py
@@ -1,56 +1,13 @@
 # Created by Xinyu Zhu on 10/22/2019, 10:11 PM
 
 import sys
-import datetime
 import json
+import os
+from .experiment_utils import read_data, f_score_calc, get_min_max
 
-sys.path.append("../../")
+project_path = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.append(project_path)
 from detectors.relative_entropy_detector import RelativeEntropyDetector
-
-
-def to_timestamp(timestr):
-    if timestr.endswith(".000000"):
-        timestr = timestr[0:-7]
-    timestamp = datetime.datetime.strptime(timestr, '%Y-%m-%d %H:%M:%S').timestamp()
-    return timestamp
-
-
-def read_data():
-    file_path = "../../../data/NAB_data/labels/combined_windows.json"
-    with open(file_path) as f:
-        content = f.read()
-        obj = json.loads(content)
-    result = {}
-    for key in obj:
-        result[key] = []
-        for slot in obj[key]:
-            result[key].append([to_timestamp(slot[0]), to_timestamp(slot[1])])
-
-    data = {}
-    for key in result:
-        path = "../../../data/NAB_data/data/" + key
-        data[key] = []
-        with open(path) as f:
-            content = f.read().split("\n")[1:-1]
-            for line in content:
-                line = line.split(",")
-                timestamp = to_timestamp(line[0])
-                obj = {"timestamp": timestamp, "value": float(line[1])}
-                label = 0
-                for slot in result[key]:
-                    if slot[0] <= timestamp <= slot[1]:
-                        label = 1
-                obj["label"] = label
-                data[key].append(obj)
-
-    """
-    data = 
-    {
-        "path":[{"timestamp":time, "value":value, "label":1/0}],
-    }
-    """
-    return data
-
 
 if __name__ == '__main__':
     data = read_data()
@@ -74,10 +31,6 @@ if __name__ == '__main__':
     all_data_keys = list(data.keys())
     all_data_keys.sort()
 
-    all_data_keys = all_data_keys[0:]  # 57
-
-    # detector = HTMAnomalyDetector("timestamp", "value")
-
     result_collector = []
     for window in window_size:
         for nbin in n_nins:
@@ -88,13 +41,7 @@ if __name__ == '__main__':
                     result[key] = []
 
                 data_value = data[key]
-                min_value = 10e10
-                max_value = -10e10
-                for i in range(min(probation_number, len(data_value))):
-                    if min_value > data_value[i]["value"]:
-                        min_value = data_value[i]["value"]
-                    if max_value < data_value[i]["value"]:
-                        max_value = data_value[i]["value"]
+                min_value, max_value = get_min_max(data_value, probation_number)
 
                 detector.initialize(input_min=min_value, input_max=max_value, n_bins=nbin, window_size=window)
 
@@ -120,37 +67,7 @@ if __name__ == '__main__':
                     if data_value[i]["label"] == 1:
                         windowed_result[window_index]["window_label"] = 1
 
-                true_positive = 0
-                true_negative = 0
-                false_positive = 0
-                false_negative = 0
-
-                for i in windowed_result:
-                    if windowed_result[i]["window_label"] == 1:
-                        if windowed_result[i]["window_result"] == 1:
-                            true_positive += 1
-                        else:
-                            false_negative += 1
-                    else:
-                        if windowed_result[i]["window_result"] == 1:
-                            false_positive += 1
-                        else:
-                            true_negative += 1
-
-                if true_positive + false_positive != 0:
-                    precision = true_positive / (true_positive + false_positive)
-                else:
-                    precision = 1
-
-                if true_positive + false_negative != 0:
-                    recall = true_positive / (true_positive + false_negative)
-                else:
-                    recall = 1
-
-                if precision + recall != 0:
-                    f_score = 2 * (precision * recall) / (precision + recall)
-                else:
-                    f_score = 0
+                f_score, precision, recall = f_score_calc(windowed_result)
 
                 result[key].append({
                     "window": window, "F": f_score, "precision": precision, "recall": recall, "n_bins": nbin
